fix: remove deps on custom.d
