+++ packages/client/types/AuthToken.d.ts
@@ -1,7 +1,7 @@
 export interface AuthToken {
   sub: string
   tms: string[]
-  rol?: 'su' | 'impersonate'
+  rol?: 'su' | 'impersonate' | null
   bet?: 1
   iat: number
   iss: string
