+++ packages/server/types/GQLRequest.d.ts
@@ -1,0 +1,18 @@
+import type {AuthToken} from '../../client/types/AuthToken'
+
+export interface GQLRequest {
+  authToken: AuthToken
+  ip?: string
+  socketId?: string
+  variables?: {[key: string]: any}
+  docId?: string
+  query?: string
+  rootValue?: {[key: string]: any}
+  dataLoaderId?: string
+  // true if the query is on the private schema
+  isPrivate?: boolean
+  // true if the query is ad-hoc (e.g. GraphiQL, CLI)
+  isAdHoc?: boolean
+  // Datadog opentracing span of the calling server
+  carrier?: any
+}
