+++ packages/server/types/custom.d.ts
@@ -1,7 +1,6 @@
 import type {GraphQLSchema} from 'graphql'
 import type nestGitHubEndpoint from 'nest-graphql-endpoint/lib/nestGitHubEndpoint'
 import '../../client/types/reactHTML4'
-import type AuthToken from '../database/types/AuthToken'
 import type ScheduledJobMeetingStageTimeLimit from '../database/types/ScheduledJobMetingStageTimeLimit'
 import type ScheduledTeamLimitsJob from '../database/types/ScheduledTeamLimitsJob'
 export interface OAuth2Success {
@@ -23,23 +22,6 @@ export interface OAuth2Error {
   error_description?: string
   error_uri?: string
 }
-export interface GQLRequest {
-  authToken: AuthToken
-  ip?: string
-  socketId?: string
-  variables?: {[key: string]: any}
-  docId?: string
-  query?: string
-  rootValue?: {[key: string]: any}
-  dataLoaderId?: string
-  // true if the query is on the private schema
-  isPrivate?: boolean
-  // true if the query is ad-hoc (e.g. GraphiQL, CLI)
-  isAdHoc?: boolean
-  // Datadog opentracing span of the calling server
-  carrier?: any
-}
-
 export type ScheduledJobUnion = ScheduledJobMeetingStageTimeLimit | ScheduledTeamLimitsJob
 
 export type RootSchema = GraphQLSchema & {
