+++ packages/server/utils/PubSubPromise.ts
@@ -1,7 +1,7 @@
 import type {ExecutionResult} from 'graphql'
 import ms from 'ms'
 import GQLExecutorChannelId from '../../client/shared/gqlIds/GQLExecutorChannelId'
-import type {GQLRequest} from '../types/custom'
+import type {GQLRequest} from '../types/GQLRequest'
 import {Logger} from './Logger'
 import RedisInstance from './RedisInstance'
 import {getUserId} from './authorization'
